add MyShapeCreator Struct and movee to InputView's function for create MyShape type method
