+++ CoordinateCalculator/UnitTestCoordinateCalculator/UnitTestCoordinateCalculator.swift
@@ -9,10 +9,10 @@
 import XCTest
 @testable import CoordinateCalculator
 class UnitTestCoordinateCalculator: XCTestCase {
-    var inputViewTest1 = InputView()
-    var inputViewTest2 = InputView()
-    var inputViewTest3 = InputView()
-    var inputViewTest4 = InputView()
+    var myShapeCreatorTest1 = MyShapeCreator()
+    var myShapeCreatorTest2 = MyShapeCreator()
+    var myShapeCreatorTest3 = MyShapeCreator()
+    var myShapeCreatorTest4 = MyShapeCreator()
 
     override func setUp() {
         super.setUp()
@@ -23,14 +23,14 @@ class UnitTestCoordinateCalculator: XCTestCase {
     }
     
     func testCheckingPrintMyPointPosition() {
-        let myValue = try! inputViewTest1.checkCountOfInputValue(inputValue: "(10,10)")
+        let myValue = try! myShapeCreatorTest1.checkCountOfInputValue(inputValue: "(10,10)")
         let myShape = myValue.calculateOfPosition()
         XCTAssertEqual(10, myShape[0].x)
         XCTAssertEqual(10, myShape[0].y)
     }
     
     func testMyLineCheckError() {
-        let myValue = try! inputViewTest2.checkCountOfInputValue(inputValue: "(10,10)-(14,15)")
+        let myValue = try! myShapeCreatorTest2.checkCountOfInputValue(inputValue: "(10,10)-(14,15)")
         let myShape = myValue.calculateOfPosition()
         XCTAssertEqual(10, myShape[0].x)
         XCTAssertEqual(10, myShape[0].y)
@@ -39,7 +39,7 @@ class UnitTestCoordinateCalculator: XCTestCase {
     }
     
     func testMyTriangleCheckError() {
-        let myValue = try! inputViewTest3.checkCountOfInputValue(inputValue: "(10,10)-(14,15)-(20,8)")
+        let myValue = try! myShapeCreatorTest3.checkCountOfInputValue(inputValue: "(10,10)-(14,15)-(20,8)")
         let myShape = myValue.calculateOfPosition()
         XCTAssertEqual(10, myShape[0].x)
         XCTAssertEqual(10, myShape[0].y)
@@ -50,7 +50,7 @@ class UnitTestCoordinateCalculator: XCTestCase {
     }
     
     func testMyRectCheckError() {
-        let myValue = try! inputViewTest4.checkCountOfInputValue(inputValue: "(10,10)-(22,10)-(22,18)-(10,18)")
+        let myValue = try! myShapeCreatorTest4.checkCountOfInputValue(inputValue: "(10,10)-(22,10)-(22,18)-(10,18)")
         let myShape = myValue.calculateOfPosition()
         XCTAssertEqual(10, myShape[0].x)
         XCTAssertEqual(10, myShape[0].y)
